Made updates necessary to write queries via the database
