+++ query_results.txt
@@ -1,0 +1,12 @@
+# SQL Query Results
+Query from client 405568: select * from users;
+
+id = 1
+name = Alice
+age = 30
+id = 2
+name = Bob
+age = 25
+id = 3
+name = Charlie
+age = 35
