--- server.c
@@ -1,198 +1,0 @@
-#include "domain_socket.h"
-
-#include <stdlib.h>
-#include <sys/stat.h>
-#include <sys/types.h>
-#include <errno.h>
-#include <unistd.h>
-#include <stdio.h>
-#include <string.h>
-#include <sys/wait.h>
-#include <fcntl.h>
-#include <poll.h>
-#include <signal.h>
-#include <stdbool.h>
-
-#define MAX_BUF_SZ 128
-#define MAX_CLIENTS 8
-#define MAX_FDS 16
-
-void unlink_domain_socket(int status, void *filename)
-{
-    unlink(filename);
-    free(filename);
-}
-
-#define MAX_BUF_SZ 128
-#define MAX_FDS 16
-
-void sigint_handler(int sig)
-{
-    // Perform cleanup or other actions before exiting
-    exit(0); // Exit the program
-}
-
-void server(int num_clients, char *filename)
-{
-    char buf[MAX_BUF_SZ];
-    struct pollfd poll_fds[MAX_FDS];
-    int new_client, amnt, i, socket_desc;
-    struct client client_list[MAX_CLIENTS];
-
-    for (int j = 0; j < MAX_CLIENTS; ++j)
-    {
-        client_list[j].domain_socket = -5;
-        client_list[j].pid = -5;
-    }
-
-    socket_desc = domain_socket_server_create(filename);
-    int num_fds = 0;
-
-    /* should do proper cleanup */
-    if (socket_desc < 0)
-    {
-        exit(EXIT_FAILURE);
-    }
-    on_exit(unlink_domain_socket, strdup(filename));
-
-    // init the poll fds
-    memset(poll_fds, 0, sizeof(struct pollfd) * MAX_FDS);
-
-    // the service is itself is a file descriptor, so add to the table of fds
-    poll_fds[num_fds++] = (struct pollfd){
-        .fd = socket_desc,
-        .events = POLLIN,
-    };
-
-    printf("SQL Server Starting!\n");
-    for (;;)
-    {
-
-        printf("Polling...\n");
-        int ret = poll(poll_fds, num_fds, -1);
-        if (ret == -1)
-            exit(EXIT_FAILURE);
-
-        printf("Polled something\n");
-
-        // accept a new client
-        if (poll_fds[0].revents & POLLIN)
-        {
-
-            if ((new_client = accept(socket_desc, NULL, NULL)) == -1)
-            {
-                exit(EXIT_FAILURE);
-            }
-
-            // Set up our client struct in an available index
-            for (int j = 0; j < MAX_CLIENTS; ++j)
-            {
-                if (client_list[j].pid == -5)
-                {
-                    printf("j = %d\n", j);
-                    client_list[j].domain_socket = new_client; // Sets our FD to the socket FD
-                    fflush(stdout);
-                    client_list[j].pid = rec_pid(new_client, buf, MAX_BUF_SZ - 1); // Get the PID from the client
-                    break;                                                         // Move on, we've already got a spot for the new client
-                }
-                else if (j == MAX_CLIENTS - 1)
-                { // No client spots available, kill the client connection
-                    close(new_client);
-                }
-            }
-            // Client set up complete, now we can do our work
-            printf("Client setup complete\n");
-            fflush(stdout);
-
-            // add it to the polling list
-            poll_fds[num_fds++] = (struct pollfd){
-                .fd = new_client,
-                .events = POLLIN};
-            poll_fds[0].revents = 0;
-
-            // successful connection
-            printf("Server: New client connected with new file descriptor %d.\n", new_client);
-            fflush(stdout);
-        }
-
-        // Event loop that handles when new clients are connected
-        for (i = 1; i < num_fds; i++)
-        {
-            /* Done with communication with this client */
-            // Now we need to remove the client from our active client list
-            for (int j = 0; j < MAX_CLIENTS; ++j)
-            {
-                if (client_list[j].domain_socket == new_client)
-                {
-                    if (client_list[j].pid % 2 == 0)
-                    {
-                        // Open file to send back to client
-                        printf("Opening private queries...\n");
-                        int fd = open("query_results_private.txt", O_RDONLY);
-                        if (fd == -1)
-                        {
-                            perror("open");
-                            exit(EXIT_FAILURE);
-                        }
-                        send_fd(new_client, fd);
-                        close(fd);
-                    }
-                    else
-                    {
-                        // Open file to send back to client
-                        printf("Opening public queries...\n");
-                        int fd = open("query_results_public.txt", O_RDONLY);
-                        if (fd == -1)
-                        {
-                            perror("open");
-                            exit(EXIT_FAILURE);
-                        }
-                        send_fd(new_client, fd);
-                        close(fd);
-                    }
-                    client_list[j].pid = -5; // Set slot open for new clients
-                    client_list[j].domain_socket = -5;
-                    break;
-                }
-            }
-            // Client removed from active client list
-            close(new_client);
-
-            // Now we close the client connection descriptor
-            printf("server: closing client connection %d\n", poll_fds[i].fd);
-            close(poll_fds[i].fd);
-            poll_fds[i] = poll_fds[num_fds - 1];
-            num_fds--;
-            i--;
-        }
-    }
-    close(socket_desc);
-    exit(EXIT_SUCCESS);
-}
-
-int main(void)
-{
-    char *channel_name = "db_server";
-    int nclients = 4;
-    int i;
-
-    // Signal Cleanup Handling
-    if (signal(SIGINT, sigint_handler) == SIG_ERR)
-    {
-        perror("signal");
-        return 1;
-    }
-
-    // start the
-    if (fork() == 0)
-    {
-        server(nclients, channel_name);
-    }
-    /* wait for the server to create the domain socket */
-
-    /* wait for all of the children */
-    while (wait(NULL) != -1)
-        ;
-
-    return 0;
-}
